Track 'last continued thread' and 'last query thread' explicitly.
The goal is to avoid depending on ReplaySession::current_task in
GdbServer, since that doesn't always match up with the thread gdb is
interested in, especially when we're executing code from task-groups that
aren't being debugged.

Generally avoiding passing Task* around is also a good idea because
they can be created and destroyed while doing operations on the
timeline.