Stop gdb immediately on the first continue-execution command after a "run".
This means that "restart" and "run" commands stop immediately after the
restart has completed, instead of carrying on. This matches gdb's
behavior for its built-in "restart". It doesn't match gdb's behavior
for built-in "run", but it seems more useful, especially when using
"run NNN" to restart at a particular event.

Resolves #1504.