Fix libpthread patching and add a test for it.
Sometimes libpthread is called "libpthread.so" but sometimes it's
"libpthread-2.21.so" (for example). So we need to generalize the file
name check. We also need to avoid assuming the .data section's
physical location in the file is the same as its logical address.
And we need to handle cases where a symbol's location is in the
virtual address range that was mapped, but is not accessible there.