Make ctrl-C interrupts more robust by continuing replay until we reach a task in the debuggee task group.
This patch also does some reformatting due to a clang-format update.

It also switches a lot of places to stop with signal 0 instead of SIGTRAP,
which is a slightly nicer debugging experience.

It also means that after resuming a checkpoint where there's a breakpoint
at the current IP, continuing execution will not stop at that breakpoint.
That seems like better behavior than the current behavior where we do
stop at that breakpoint.