Clean up signal handling, especially during replay.
When we try to deliver a signal, sometimes a new SIGSEGV is raised instead.
This happens when there isn't enough stack space to allocate a signal handling
frame. In this case we want to record three events in the rr trace: raising
the original signal (EV_SIGNAL), raising SIGSEGV (EV_SIGNAL), and then
delivering SIGSEGV (EV_SIGNAL_DELIVERY).

To replay this correctly we need to separate handling of EV_SIGNAL and
EV_SIGNAL_DELIVERY. Prior to this patch handling EV_SIGNAL would automatically
transition into signal delivery, but now we have a raised signal which
is not delivered. This makes handling signals more consistent with all other
kinds of events, where current_step.action is initialized in
setup_replay_one_trace_frame.