Remove timeslice coaelescing optimization.
This causes problems in ReplayTimeline::run_forward_to_intermediate_point,
which isn't aware of this optimization. In general it's much simpler for
consumers of ReplaySession to assume that events are not coalesced, so
when we have a series of SCHED events but hit a breakpoint before the
first one, the reported event number really is the number of the first
SCHED event, not the last one.

Testing this optimization on the chew_cpu test, replay times were
barely affected. Using a smaller context switch interval would
increase the overhead of disabling this optimization, but would also
necessarily increase recording overhead. Disabling this optimization
at worst makes replay context switches no more expensive than the
same context switches during recording.