When resetting the syscallbuf, zero out its data.
We maintain the invariant that data outside the currently-used area of the
syscallbuf is zero. This means that when creating a checkpoint during replay,
copying only the currently-used area of the syscallbuf is enough to ensure
that the entire syscallbuf region is consistent between the original state
and the checkpoint. Thus, stray writes performed by return_addresses()
and possibly the tracee itself are more likely to be consistent if they
land in the syscallbuf.