Remove cpu_set_t from kernel_abi.h
sizeof(cpu_set_t) is not actually kernel ABI. The only place we're using it is
sched_getaffinity, which takes a length argument telling us the size of the
output buffer, and returns the amount of data written (the system call,
that is, not the glibc wrapper). So modify the recording of sched_getaffinity
to use those values and make no assumptions about sizeof(cpu_set_t).

Resolves #1549.