Rework set_thread_area handling.
Instead of being just one thread-area, there are actually multiple entries,
one per entry number, so store a vector of thread areas instead of just
zero or one.

Also, when clone()ing with CLONE_TLS, on x86-64, the parameter is not
a thread_area so don't treat it as one. This didn't really cause
any problems but it was very confusing.