Fix bug where we assumed the ticks-count for a DESCHED event was correct when it isn't.
The tick count for a DESCHED-arm event is not reliable, since during
recording that event is "made up" when we discover that a system call has
been interrupted. I.e. its tick count does not correspond to the tick count
when the ioctl was executed.

This created a problem during replay. In ReplaySession::try_one_trace_step,
if the requested max tick count is less than the tick-count for the current
event, we run until that tick-count is reached, assuming that that will
happen before we reach the event (in this case, a DESCHED ioctl). But of
course this is incorrect when the tick count for a DESCHED event is greater
than the tick count when its ioctl was performed.

The fix is to simply not use this "fast path" for DESCHED arms. Instead, rely on
cont_syscall_boundary to stop us before we reach the ticks target.