Clean up syscallbuf resetting a little bit.
It's safe to reset the syscallbuf after each time we've recorded some other
event.

This also improves our assertions to detect during recording when we've
flushed a syscallbuffer and then another complete syscall is recorded.
(Replay can't handle that.)

This patch also aligns the clearing of the syscallbuf during recording
with the recording of the syscallbuf reset event. This ensures that
clearing happens at the same relative time during recording and replay.