Make syscallbuf slightly more efficient by using privileged untraced syscalls in a few more places.
When a syscall result is/should be zero and it has no user-space side effects,
we can use the privileged untraced syscallpath so that during replay we don't
even enter the kernel, just clear the syscall result register. During recording
we don't have to trap to rr or record the syscall state.