Restore syscall buffer contents all at once instead of syscall by syscall
This requires making sure the syscallbuf code does not corrupt the syscall
buffer contents during replay. In particular in/out parameters must not
be copied to the syscall buffer before the syscall, during replay.

We also now get the syscall result from the syscall buffer in the
syscallbuf code, rather than having rr set it.