Make breakpoint_overlap test more robust by generating a lot more conditional branches
I observed an intermittent test failure where breakpoint_overlap triggered a
SCHED event with 1014 conditional branches elapsed since the getgid32() call.
(The test is recorded with a nominal 100-ticks-per-timeslice limit.)
This triggered a failure because we'd finished the main part of the test and
were deep in libc code preparing to exit, and we couldn't set a breakpoint
there before libc was loaded. We need the SCHED event to fire inside the main
part of the test, so let's just add a lot more conditional branches to the
main part of the test to catch this much delay in firing the interrupt
(and more!).