Improve chaos mode
The big change here is to revise the scheduling of high-priority-only
intervals.

Another important change is that we interrupt the tracee when a
high-priority-only interval is about to begin or we need to refresh
chaos mode settings. This lets us start a high-priority-only interval
as soon as we reach the correct time, instead of having to wait
for the tracee to yield --- very important when using very
frequent, short high-priority-only-intervals.