use configure_file() to put files in the binary directory
CMake will automatically care for the dependency and regenerate the file if the
input changes, as well as doing the dependency tracking that the input file is
there before the target is built.

While at it this fixes a lot of places where building would only work if the
source and build dir are the same as the properties and information about
custom command output were relative to the source directory because they had
no explicit path set.