Update logging code so logging is always compiled in
Having to rebuild to turn on debug logging is annoying. Users should be
able to turn on logging without rebuilding. We also need APIs to
dynamically turn on logging from the debugger or by hacking rr to
turn logging on at interesting times.

This approach obviates the need for DEBUGTAG entirely. We use __FILE__
to generate the module names instead. The next patch will remove DEBUGTAGs.