Don't close stderr in unshare test
I found an issue that caused this test program to fail, even when not
running under rr.  Leaving stderr open allowed the assertion failures
from the child to show up in the console, which made it possible to
debug.

Maybe it's useful to leave this in to make debugging future problems
easier, unless there's some downside that I'm not aware of.