Fix set_tid_address when child threads run first
See https://github.com/mozilla/rr/issues/1680 for full discussion.
On my machine, the child thread running run_thread would execute before
pthread_create returned in the parent thread.  This would cause v to be
set to 0 and futex to fail with EAGAIN.  A normal program would just
retry, but we want this test to verify that the terminating thread
triggers a FUTEX_WAKE.  Therefore, use a pipe to delay the child thread
until (hopefully) we enter futex in the parent.