Refactor syscall hooking
Without this, the preload library had a collection of pre-layed-out stubs
to which it added CFI, which would then be overwritten by rr once it knew
the actual addresses it needed. However, in addition to this, we also needed
"extender pages" which are within 2GB of the patches callsite in order to
make use of a relative jmp. This commit removes one layer of indirection
by merging the stubcode into the extender page and using CFI in the actual
syscall hook to ensure that backtraces are still correct.