Allow moving the ASSERT ostream out of hot path
Constructing the ostream (and stringifying any logged arguments) is
moderately expensive. This work only needs to be done if the assertion
is triggered, so provide a second macro that behaves that way.
Uses not in the hot path (in my application) are not renamed for
now to avoid unnecessary code churn.