Make exe_directory work when rr is used as a library
This is done by looking through the mappings in the current address space
and finding the one that contains rr code, under the assumption that
that would be either the rr executable or the rr library.

This is done by reuising the code from AddressSpace to iterate over kernel
mappings.