Reimplement vfork() support
We were implementing vfork() as just fork(), but this breaks some
applications that expect the vfork child to share address space with
the vfork parent.

During recording, we perform a real vfork(). The kernel ensures that
the vfork parent does not return from the vfork() syscall until the
vfork child has execed or exited, and the scheduler observes that; we
don't need any changes to our scheduler.

During replay, we can't perform a real vfork() because the parent must
be runnable so we can exit the syscall immediately (preserving the replay
invariant that replay tasks are not in the kernel except while performing
specific syscalls on behalf of rr). So instead we do a clone() with
CLONE_VM, i.e. a fork() but with the address space shared between parent
and child.

Resolves #1741