Disable grsecurity MPROTECT on rr and its exec_stubs
We need to do this because we need to be able to write to code in the
spawned tracee processes forked from rr itself and execed with the
exec_stubs.

We also need to be able to write to code in the recorded tracee
processes, but users are responsible for disabling MPROTECT for those
themselves.