Rework handling of the signals-race-with-exit problem
We still have intermittent test failures caused by signals
(usually TIME_SLICE_SIGNALs) being stashed for a task just before it
reaches an exit()/exit_group() syscall. This triggers an assertion
failure because once we've committed to exit, such signals are going
to be dropped. In general it would be wrong for an exiting task to
accept a process-wide signal and then just drop it; outside of rr,
the task either accepts the signal and runs a signal handler before
exiting, or exits and another task gets a chance to accept the signal.
These test failures usually seem to occur because some signal gets
received while we're in syscallbuf code leading up to the exit syscall,
so we stash it until we're outside the syscallbuf, but that never
happens because we exited. A similar bug could occur if the tracee
does a sigprocmask to block a signal which has been stashed; the signal
will be delayed indefinitely, whereas outside of rr the kernel would
have chosen a different task to receive the signal if there is a task
which has the signal unblocked.

Here we're trying to fix this issue for good by creating an invariant
that we never proceed with an exit/sigprocmask syscall while there are
stashed signals. In RecordSession, if we've entered one of these syscalls
and there are stashed signals, we pretend that the syscall is interrupted
by a signal. We change the syscall number to something innocuous to get us
out of the kernel and then set register state so it looks like an
interrupted syscall. We are then able to process any stashed signals like
we would if they interrupted a normal blocking syscall. After that,
we force a restart of the syscall by explicitly triggering a
restart_syscall().

This means a signal handler could theoretically interrupt an exit()
and cancel it by sigreturn()ing somewhere else. That's weird but
shouldn't be a big deal in practice since outside rr it could just
arrive just before entering the exit() syscall and get the same
effect.