Only consider files when looking up executables in $PATH
dc23125f make `rr record` do $PATH lookup on its own to check the
executable, but only checks if a matching file has the executable bit,
which is true of directories. In some corner case, where a directory of
the same name as a program is in one of the directories in $PATH, this
makes `rr record` fail with:
    execve of '...' failed (EACCES)

So first check if a matching file is a regular file.