Don't clobber memory during execve if we vforked
`process_execve` happens during the return from execve, where at record
time the old address space was already destroyed and the parent was
continued. Thus, clobbering the stack here may be oberservable to the
parent during replay. Avoid this problem by checking if there are other
thread groups that share this address space and restoring the memory if
so.