[AutoRemoteSyscalls] Add mmap fallback when no stack is found
When replaying an rr replay, the inner tracee can end up in a a situation where
it has no mapping that can be used as AutoRemoteSyscall's scratch area. If this
happens while we are trying to clone that task (e.g. during reverse stepping),
we used to crash because cloning (or rather opening the mem fd afterwards)
requires memory parameters. To mitigate this add a worst-case fallback, which
allocates one page worth of RW memory to use as a scratch area.