Don't deliver synthetic SIGCHLDs until the syscall completes.
The kernel doesn't interrupt wait/etc to deliver ptrace signals. Instead the
syscall finishes successfully and only then does the kernel deliver the signal.
This is visible by strace-ing ptrace_trace_clone. We can do this by simply
deferring the synthetic signal when we're in a syscall restart and restarting
the syscall. The synthetic signal will automatically be delivered when the
restarted syscall finishes.