Fix handling of multiple signals without handlers and syscall interruption.
The kernel's signal processing algorithm for interrupted syscalls is roughly:

  while (true) {
    // Do boring work
    if (has_actionable_signals) {
      while (true) {
        if (!dequeue_signal())
	  break;
	ptrace_signal();
	if (handler_is_ignore)
	  continue;
	// Otherwise invoke the signal handler
      }
      ip -= 2;
      ax = syscall_to_restart;
      restore_saved_sigmask();
    }
    // Do more boring work
    if (done)
      break;
  }
  enter_userspace();

Consider a case where a ppoll blocks signal A, and is interrupted by signals B
and C. All three signals are SIG_IGN. The kernel will deliver signals B and C,
only then reverting ppoll's effects on the sigmask and unblocking signal A.
Before it does this, it will set up the syscall restart by adjusting ip/ax.
Then we go around the outer loop once more, delivering the newly unblocked
signal A, before finally returning to userspace.

This commit makes rr correctly handle cases that trigger multiple iterations of
either the inner or outer loops.  We do not restore the sigmask saved by
ppoll/pselect6 until the inner loop has finished, and we record the kernel's
adjustments to ip/ax before performing a second iteration of the outer loop.
This ensures that we deliver signal A at the appropriate time, that our signal
states match the kernel's at all times, and that signal A can be delivered
during replay because the modifications to ip/ax are now recorded.

Fixes #1830.