Don't link default startfiles or libs with rr_exec_stub
This ensures that the first_instruction test passes even if symbols
for ld.so aren't found, because the first instruction will be in rr_exec_stub
for which we definitely build symbols. For some reason in Ubuntu 16.06,
we don't find symbols for ld.so if the tests are run via a command
launched by ssh. (They are found if the tests are launched from an ssh
shell.)