Fix a crash while replaying `rr record rr replay`
Similar to 55e68a94ab37ec690b78a1443607b2a37560d486, when recording an rr replay,
a new tracee will be stripped of all of its mappings, except for the rr page,
after which the mappings are put back by the tracee rr.
After the recent thread_locals change that includes the thread_locals page.
We already checked for the situation where that page was unmapped. However,
we did not cover the situation where the page was unmapped and then replaced
by something else, which caused rr to crash during replay of the replay.