Make exit_syscall robust against which ptrace trap we're on
With the new seccomp ordering, we're not consistent about where we
process the syscall entry. `exit_syscall` was assuming that the next
trap we hit would be the syscall exit trap (in the absence of signals).
However, if we're at the entry trap, but before the seccomp trap,
that assumption is not true. Detect that situation and allow it.

An alternative implementation would be to try to always move forward
to the seccomp trap, but that seems signficiantly more complicated.

Fixes #1851