Don't lose signal mask manipulations that happen right before exec
We were buffering the signal mask in the syscallbuf_hdr, but neglecting
to reload it before throwing it away during the exec, causing a
divergence between rr's and the kernel's understanding of the signal mask.
This was triggered on Ubuntu 16.10 during the ptrace_remote_unmap test,
because `raise` now calls sigprocmask (and in that test we call raise
right before exec). However, I still added a test for this specific problem
to avoid depending on libc implementation details.