Improve Monkeypatcher interfering branch heuristic
On Ubuntu 16.10, the glibc read syscall wrapper, is not patched, due to
what rr thinks is an interfering branch. In reality, it's just an unfortunate
subsequence of a different instruction in a different function. We can
fix this (and the severe performance regression caused by it) by improving
the heuristic, since this unfortunate byte sequence appears to target,
one byte after the end of the syscall instruction, which is not a valid
jump target for an actual interfering branch.