Support `write` calls on monitored file descriptors
We used to only support pwrite on `/proc/*/mem` and any fd, whose backing
file was mapped MAP_SHARED into a Task's address space. For a
regular `write` system call, we can't easily get the offset to write at
by just looking at the register state (which we need to know in order
to reflect these changes appropriately into memory). Luckily, the kernel
allows us to query a file descriptor's current position by reading
/proc/*/fdinfo. Use that to get the offset, but take care to only read
it when we're actually going to need it.