Add register data to PATCH_SYSCALL events
Prior to this change, PATCH_SYSCALL is the only case where register values
can be changed via something other than user-space instruction execution
without the new values being explicity recorded in the trace. Replay worked
because we actually entered the kernel, which performed the correct effects.

Recording registers here means that those effects are saved in the trace and
we don't have to depend on the kernel performing them consistently. This is
useful for non-replay consumers of the trace. The performance overhead should
be minimal because PATCH_SYSCALL happens at most once per occurrence of a
syscall instruction.