Allow replay of executables that end in `.so`
Such as the dynamic linker, which can be run as an executable in order to test
changes to the dynamic linker itself. If a dynamic linker is present, that's
where the ip will be right after the execve, so in order to find the main
executable, we simply have to find a file that does have an executable
segment, but of which no segment contains the ip. If there is only one file
with executable segments, the file is likely statically linked (it is probably
theoretically possible for an executable to be its own dynamic linker, thus
putting two copies of the same file in memory, but even that should work,
since the only use of finding the main executable is for GDB and naming
the rr threads).

Fixes #1880