Properly handle a ptracer changing a ptracee's syscallno
First, we can't go through the syscallbuf code, since the ptracer may change
the syscallno to an arbitrary different syscall, which the syscallbuf code is
not designed to handle. To achieve this, set bit 2 in syscallbuf_hdr->locked
whenever the ptracer is observing syscall entries (thus forcing the syscallbuf
code to do a regular traced syscall and not make assumptions on what happened).
Additionally, fix a bug where the event's syscallno was not updated when it
was changed by the ptracer.