Fix replay of syscall instructions on MAP_SHARED pages
When PTRACE_SYSEMU, causes a syscall-entry trap, this trap is functionally equivalent
to such a trap caused by PTRACE_SYSCALL, except that the syscall will definitely not
be executed. In particular, this means that, if execution is resumed with PTRACE_SYSCALL,
we will see a syscall-exit trap. Thus, in order to be safe, we need to get out of the
kernel before moving on. `finish_emulated_syscall` is supposed to do this. However,
the way it was previously implemented, it could require a remote syscall in order
to set a breakpoint. Unfortunately, setting such a breakpoint requires executing
a remote syscall, which uses PTRACE_SYSCALL, causing us to see the previous exit
trap, instead of the new entry trap, confusing everything. Luckily, it appears
that we can just rewrite `finish_emulated_syscall` to use PTRACE_SYSCALL, which
will just go to the exit trap. There was one call to this function which appeared
to be incorrect, but with that removed, all tests pass without problem, and
syscalls on MAP_SHARED pages no longer present a problem.