Fix PTRACE_SYSEMU entry stop with invalid syscall number
This test was intended to make sure we were correctly emulating PTRACE_SYSEMU
behavior I'm relying on in the next commit. However, the test instead
uncovered a different issue in ptrace emulation, which is that we
used to skip the syscall-entry stop if the orignal syscall number going
into the syscall is negative. Fix that, and keep the test, which now
tests both issues.