Fix rare corner cases in timer expiration
Issue 1:
If a timer expires in userspace the orig_rax of the corresponding
ptrace trap will be negative, reflecting the received IRQ. We faithfully
record this. Usually it doesn't show up as a problem, because we single
step (or otherwise use traps to) step our way to the point where the
async signal occurred, resetting t->regs().orig_rax to some negative
value, so orig_rax is negative in both recording and replay and everything
is fine. However, if we get unlucky and the timer expires right after
we return from a system call, without any intervening instructions,
we don't need to do any stepping during replay, so orig_rax, won't
become negative (retaining whatever value it had at syscall exit).
Thus, since the value of orig_rax is still positive during replay
(and negative in the recording) we do perform the comparison, find
a mismatch and ignore the correct execution point. Fix this by
ignoring orig_rax, whenever one of the two registers is negative,
even if the other one isn't. Meaningful comparison (i.e. the registers
indicating a syscallno) should always have positive register values.

Issue 2:
This happens when the timer expires right after a sigreturn. The kernel
will or in 0x3 to the segment register's RPL, but if this is the null
selector, this would get canonicalized back to 0x0 (I believe by the CPU).
However, we do record the 0x3 in the sigreturn ptrace trap, and if we
don't return to userspace between then and when the timer expires, we
will also record the 0x3 in the signal register state. Fix this by
ignoring the RPL of the segment registers. For any non-null segment
this should be this should be 0x3 always anyway.