Allow 240s for tests on KNL
As far as I can tell, all test fit into 120s when run in isolation, but if
two tests happen to get scheduled on the same core, things can get quite
a bit slower. This is especially bad when all four hyper-threads per core
are active, since the OS doesn't take care to avoid overloading the physical
core. Raise the time limit to 480s, which should be more than sufficient.