Best effort support for cross-architecture syscalls
When the tracee uses `int $0x80` to perform a system call, the kernel
uses the 32bit system call table to execute that call. Without this
patch, we'd just assume it was a proper 64bit syscall, wreaking havoc.
This patch fixes the most obvious issues here (converts registers,
uses the correct syscall table) and works fine on a simple example
I had that made use of this "feature", but of course there's likely
a long tail of issues that would be caused by this. I have no
intention of actually making sure every syscall works in this mode,
becuase I suspect very few applications are actually relying on this.
If there are such application, there'll now be a big fat warning in
the log, and we can revisit improving this support at such time.