Support netfilters SO_SET_REPLACE socket option
Also add a test for the same. Since in general this require CAP_NET_ADMIN,
and we want to avoid modifying the actual nat tables on the system, the
test uses user namespaces. As a result, the test is a bit more complicated,
but on many system can be run as an unprivileged user.