Work around kernel bug pwriting to PROT_READ thp
Kernels that have the DirtyCOW exploit mitigation patch (i.e. 4.9
and most stable kernels by now) will hang when trying to write to
a PROT_READ page that is backed by a transparent huge page and causes
a COW (which happens rather frequently for rr when it first sets up
a new address space). A kernel patch to fix this is pending at
http://marc.info/?l=linux-kernel&m=148359462917379&w=2, but until
then, we should avoid trying to pwrite to PROT_READ pages that may
be backed by transparent huge pages.