Fix checksum_sanity test when tmp files were cloned
In ea04c71b4283f18bf20a5d20e22b70882d155e15, I added a test to checksum_sanity
to handle the case where there was a mapping entirely beyond the end of it's
underlying file. If during replay, this mapping is crated using trace data,
the replayer will automatically create a SIGBUS region and that commit adjusted
the checksumming code to recognize and ignore such regions. However, if the
mapping is created from a cloned file (e.g. because /tmp is a btrfs file system),
we do not create a separate SIGBUS region, instead relying on the regular
semantics for mapping files to apply. The above commit caused a problem
for this case, because we wouldn't record the checksum for the mapping beyond
the end of the underlying file, expecting it to be the special SIGBUS region
during replay. However, it could just be a regular mapping instead if the
data was cloned. Handle this case, by actually recording the SIGBUS
regions in the checksum file with a special checksum and checking that
during replay, we either have one of our special regions, or we get the
same EIO error trying to access it as we did during recording.