In new CPU binding code, use cpuset.effective_cpus
`cpuset.cpus` is the user configurable maximum setting for the current cgroup,
`cpuset.effective_cpus` is the actual CPUs that members of this cgroup can run
on (e.g. if the cpus are restricted by parents or siblings). However,
some older kernels (or modern kernels with appropriate configuration),
do not have `cpuset.effective_cpus`, so still fall back to the other
one if not found.