Detect and work around KVM PMU bug
KVM has a bug where an IN_TXCP counter with nonzero period will fail to
be allocated in the host, causing the guest to silently drop events. We add
code to PerfCounters to detect this bug. When the bug is present, we stop
trying to handle HLE with an IN_TXCP counter because there's no reliable way
to work around the KVM bug. Instead we count IN_TX conditional branches and
warn (and bail) if we detect some, i.e. if HLE (or RTM, though that
never works anyway) is in use.

We'll try to get a patch upstream to fix KVM. The KVM bug is only detected
on hardware that actually supports HLE/RTM.

Resolves #1952