Workaround #1957
There isn't really a need for the useless counter, unless we deschedule our own counters. We only do so, if we have the IOC_PERIOD bug.
It so happens that this restores functionality on all kernel versions I have access to, but of course that may be a coincidence. Still, better than nothing.