Tighten scheduling assertions
Ensure we don't ever schedule a task while there's a nonempty round-robin queue
unless the scheduled task is at the front of that queue.

The test `ptrace_attach_thread_running` showed that we violated this invariant
in one case: when the task at the head of the queue is in an unstable exit,
we would go ahead and schedule the next task to report a status change, even if
it wasn't at the head of the queue. Fix this by emptying the round-robin queue
before we wait for any task to report a status change. (The round-robin queue
is meant to impose fairness when a task appears to be busy-waiting, and if
all tasks are blocked or in unstable exits, it's no longer relevant.)