Don't emulate PTRACE_SYSCALL stops for privileged syscalls
They should be treated as invisible to ptracers.

Normally we disable syscall patching completely when there's
an emulated ptracer, but that doesn't always work because we may have
already-patched syscalls in the process when the ptracer attaches.
Our emulation can't be as good in this case, but we can still ignore
syscallbuf's privileged syscalls so unexpected syscalls aren't seen.