Save and restore Task::wait_status when performing a remote syscall.
But be careful not to restore it if the new wait_status is meaningful.
The call_exit tests will fail if we restore it unconditionally, because
the ptrace exit notification from the diversion session will be lost.
More generally, if a remote syscall were to trigger a signal or similar
we shouldn't squash that.

David Baron has a recording where a SCHED is being delivered to an ip
that contains a repeating string instruction (strlen). We fast forward
through the string instruction, but the remote syscalls triggered
by a kernel bug workaround (see #2005) clobber the expected SIGTRAP, and
when we continue searching for the context switch point rr asserts
because the tracee appears to be in an unexpected syscall (because its
wait_status is the remnant of a remote syscall).

Issue #2005 is going to remove the workaround, so this specific case will not
trigger this bug anymore, but lets not leave the footgun lying around.