Remove workaround for kernel THP bug
AFAIK, the fix for the kernel bug was backported a while ago to every branch
that had the bug and the number of people still using a buggy kernel should
be very small.

Our workaround makes setting breakpoints in > 2MB executables very slow
(8 context switches per breakpoint), so let's just remove it.

Resolves #2005