Use the size of the mapping, not the size of the mapped file, when replaying mmap.
Now that we propagate writes to an fd into mappings for that file we need to map
the entire range rather than just the pages (partially) covered by the file. A
tracee can later write to the fd causing it to grow into the rest of the mapping.
If we fail to set aside the proper amount of space, we assert in
MmappedFileMonitor when attempting to propagate the write.