Create emufs shared files in TMPDIR instead of /dev/shm
This means noexec mounts of /dev/shm are not a problem. We already require
that TMPDIR point to an exec-enabled mount.

In some systems /dev/shm is a tmpfs and /tmp is a regular filesystem, so
in those systems we might regress performance a bit here. Also /tmp might
fill up when /dev/shm doesn't. I don't think this are significant issues.
This makes our story simple: make sure TMPDIR or else /tmp are exec-enabled
filesystems with plenty of space, and tmpfs if that helps performance.

Resolves #2016