Don't clear wait_status when we stash a signal
We don't need to do this. Leaving the wait_status in place means
that if rr chooses to deliver the signal, it will see that the
task is already in a signal-delivery-stop that it can use. This
avoids invoking the signal-delivery-path which can cause
problems in some cases, and is less efficient.