Abandon efforts to connect sockets from arbitrary tracees and instead keep a single socket open in every tracee.
In some cases we just can't mount /proc in an arbitrary tracee, e.g. when the
tracee is chrooted to /proc/<pid>/fdinfo where no files or directories
can be created regardless of capabilities. Keeping a single socket
open in all tracees is cleaner and more efficient, and seems to work
fine in tests. (I had previously thought it would not work across
network namespace changes; hopefully I was just wrong, although it's possible
there are bugs in older kernels.)

Resolves #2185