Allow operations on fds >= 1024 to be syscall-buffered
We treat the disabled flag for fd 1023 as meaning "disable buffering for any
fd >= 1023".

When ASAN/LSAN spawn a helper process they close all fds in a particularly naive
way, calling close() on every fd up to the process' fd limit. Something in Mozilla's
test harness sets that limit to 1M, so it's essentially that we be able to
syscall-buffer those close calls.