Hash log_modules on string contents, not the value of the char* pointer.
An unordered_map<char *, ...> will hash on the value of the pointer, not the contents of the string it points to.  That's probably not what we want -- while it's likely that the compiler will cause all instances of __FILE__ in a given file to point to the same constant string, we don't want to rely on that.  Conveniently, just replacing the "const char*" with "string" will fix the issue, since all the uses get automatically converted.