Emit a nonzero exit code from the test-runner when a test fails.
This is useful when running the tests through a test harness that checks the exit code for failure rather than parsing the test output.

This change also renames leave_data to test_passed (and inverts the "y"/"n" logic), since the variable is now used for more than just choosing whether or not to retain the test data.  As part of that, I've changed the comparisons to always compare to "y", on the idea that any bug that causes a non-"y"-or-"n" value should be treated as a failure.

Tested by confirming that the results of "make check" are consistent before and after this change, on a setup with a handful of pre-existing test failures.