Add --resource-path option.
In some cases, we may need to override the default that expects these files to be located relative to the location of the mmap'ed rr executable.  This is particularly relevant when the executable is a symlink, as the symlink is effectively resolved before computing the relative paths, and this may not be the desired behavior.