Add support for breakpoints on SystemTap probes
Userspace SystemTap probe points may be associated with a semaphore,
a counter used as a guard variable to avoid the potentially expensive
execution overhead of preparing arguments for a probe point in which no
debugger is currently interested. Setting a breakpoint on such probes
presents a dilemma for rr, as producing the behaviour gdb expects when
it increments a probe semaphore during replay would create divergence.

This patch uses a brute-force approach: it uses the LD_AUDIT machinery
to scan libraries as they're loaded for STap semaphores, incrementing
them eagerly at run-time; it also notifies rr of discovered semaphores,
so that writes by gdb to these addresses during replay can be silently
suppressed.