Handle a race between exec and writes to shared memory mappings.
The execed thread gets a new address space sometime between entering the execve syscall and exiting it. During that time, other code can run, which can write to a memory mapping that's shared with the execed thread's pre-exec address space. If the execed thread has a new address space, trying to read data out of it's /proc/pid/mem fd that we have will fail.

This would be fine except that a) MmappedFileMonitor doesn't fallibly record the data and b) we'll try to reopen the mem fd which will fail because the newly execed thread has not been waited on yet. Fix both of these, and try to ensure that if there is another thread group that shared an address space with the newly execed thread group prior to the exec that we will pick up the write by looping through all tasks on the AddressSpace looking for one to succeed.