Handle invalid path/argv in exec
Prior to this patch, passing invalid (as in faulting) addresses as
the path or argv to exec would cause rr to assert because it was
itself trying to read that memory and failing. Instead, we should
go ahead and simply abort processing of the exec on such a fault,
since we know the kernel will fail it with EFAULT as soon as it
gets the chance.