Try a more robust random_device hack
As noted in #2459, the previous random_device hack doesn't work
anymore on newer libstdc++ versions. This instead modifies the
passed `token` string in place and points it at the random device
we want to use. This is approach isn't without problem either:
We're modifying a string that's annotated in the source as `const`.
I would expect it to be fairly rare for this to be a problem, but
it's certainly possible.