Adjust an assertion that invokes undefined behavior
The GdbRequest struct has padding so memcmp'ing
is undefined behavior. I see this assertion fail intermittently
when this function is triggered, only because of differences
in the padding. We could write a proper comparison function,
but since this is just an assertion that the task exists, it's probably
sufficient to assert that we're targeting the right task.