WIP/RFC: --nested=separate for multi-threaded recording of subtrees
A common situation for us is that somewhere along the line in an
rr recording, we spawn a separate process that does a whole bunch
of work, but doesn't share any shared memory with the parent process,
nor does particularly much communication. In an ideal world of course,
rr would detect this kind of situation and record the two tasks in
parallel into the same recording. That seems like a very difficult
task, so this attempts to do the next best thing by letting the user
request this behavior manually.

In particular, any nested rr recording started with
`rr record --nested=separate` will get its own RecordSession, along
with its own recorder thread (and trace compressor threads to its
own thread). Implementation wise, this recorder thread is a thread
of the original rr process (in the sense that it shares its memory
space), but is placed in its own thread group. As a result, its
wait and ptrace state is entirely separate and the two recorders
act essentially as though they are entirely separate processes
(that just happen to share an address space).

This way, I'm hoping to keep this minimally invasive, as each
session is essentially and independent rr process.

Current status: Functional, but a few things remain to be done:
1) The CPU assignment needs to be reset, otherwise there is very
little point to this.
2) Some minor cleanup still.
3) Tests
I'm hoping to get some feedback on the general approach in the
meantime.